requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
WTForms==3.1.2
PyMySQL==1.1.1
Werkzeug==3.0.3
python-dotenv==1.0.1
Pillow==10.4.0
Flask-Migrate==4.0.7
email-validator==2.2.0
cryptography==43.0.3
